requirements.txt
matplotlib==3.5.2
numpy==1.21.4
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
scipy==1.7.3
tqdm==4.64.0
Pillow==9.1.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
pytest==7.4.0
torchinfo==1.8.0
pyyaml==6.0.2